Move runUsingRuntime and runCollectOutput to run_common.go
